Festive Update
Backend Updates
- Added Presets
- Able to change banner
- Linux should be able to read "Config.json"
- Feeling a little festive (shops will generate christmas items from the 14th dec)
- 'season' is now a float... if you are hosting '12.41' don't put '12'
~ this feature is currently used with festive shops (christmas but this will slowly be pushed to everything <3)
- Shop Bundles can give you styles now (if added)
- Fixed Level and tiers on chapter 2 (only breaks with buying the battle bundle)
- Added season 19 bp (no quests though, and technically i only added the stuff till the 19.10 update)
- Improved certain files code
- Backend loads less stuff with force season (eg. bp)
- Fixed locker random backbling
- Fixed Temp bans not popping up insta (and 0 will actually remove the message live)

Launcher Updates
- Improved settings

+++ FortBackend/src/App/Routes/ADMIN/NewFolder/ContentUpdate.cs
@@ -33,7 +33,7 @@ namespace FortBackend.src.App.Routes.ADMIN.NewFolder
         public class ServerContentRequest
         {
             public bool ForcedSeason { get; set; }
-            public int Season { get; set; } = 0;
+            public float Season { get; set; } = 0;
             public int WeeklyQuests { get; set; } = 0;
             public bool ShopRotation { get; set; } 
         }
